tif_jpeg.c: fix MSVC compiler warning
Fixes MSVC 19.44.35225.0:
D:\a\gdal\gdal\frmts\gtiff\libtiff\tif_jpeg.c(2036,40): error C2220: the following warning is treated as an error [D:\a\gdal\gdal\build\frmts\gtiff\libtiff\libtiff.vcxproj]
D:\a\gdal\gdal\frmts\gtiff\libtiff\tif_jpeg.c(2036,40): warning C4334: '<<': result of 32-bit shift implicitly converted to 64 bits (was 64-bit shift intended?) [D:

+++ libtiff/tif_jpeg.c
@@ -2040,7 +2040,7 @@ static int JPEGSetupEncode(TIFF *tif)
                 if (!TIFFGetField(tif, TIFFTAG_REFERENCEBLACKWHITE, &ref))
                 {
                     float refbw[6];
-                    uint64_t top = 1UL << td->td_bitspersample;
+                    uint32_t top = 1U << td->td_bitspersample;
                     refbw[0] = 0;
                     refbw[1] = (float)(top - 1);
                     refbw[2] = (float)(top >> 1);
